fix(compile): cleanup leftover files and fix UDS/ELM327 wiring
- Remove leftover data/, ai/, models/ files that referenced the old
  OBD2Service and Room
- UdsProtocol: add CLEAR_DTC service ID (0x14) that was missing
- UdsClient: use Byte literals for negative response
- IsotpTransport: restore return true that got dropped
- Elm327Driver: use raw port.open() (no longer wrapped) and import
  UsbManager directly; remove stale UsbDevice import
- LiveDataScreen: wrap connectCurrentDevice in coroutine scope

+++ app/src/main/java/com/example/n54guru/protocol/UdsClient.kt
@@ -37,8 +37,8 @@ class UdsClient(
         if (response.isEmpty()) throw UdsException("Empty response")
 
         if (response[0] == UdsProtocol.ServiceId.NEGATIVE_RESPONSE) {
-            val nrc = if (response.size >= 3) response[2] else 0
-            val originalSid = if (response.size >= 2) response[1] else 0
+            val nrc = if (response.size >= 3) response[2] else 0.toByte()
+            val originalSid = if (response.size >= 2) response[1] else 0.toByte()
             throw NegativeResponseException(originalSid, nrc, nrcToString(nrc))
         }
